Add thread error handling to stacktracer
stacktracer would fail if the output directory didn't exist, and without
error handling this would be lost in the logs.

Also:
- Move thread_safety_net to a new library (thread.py), as otherwise
  there was a circular import failure.
- Add a trace() function to lib.util
